restructure and fix bug in sort
