+++ profile/index.html
@@ -110,15 +110,13 @@
               tabindex="0"
             >
               <h3 class="my-3">All Users</h3>
-              <form class="mb-3" role="search">
-                <input
-                  class="form-control"
-                  type="search"
-                  placeholder="Search User"
-                  aria-label="Search"
-                  id="search-user"
-                />
-              </form>
+              <input
+                class="form-control mb-3"
+                type="search"
+                placeholder="Search User"
+                aria-label="Search"
+                id="search-user"
+              />
               <div class="list-group w-auto" id="profiles-container"></div>
             </div>
             <div
