+++ src/js/api/index.mjs
@@ -1,4 +1,5 @@
 export * from "./makeOptions.mjs";
-export * from "./auth/registerUser.mjs";
-export * from "./auth/loginUser.mjs";
 export * from "./constants.mjs";
+export * from "./posts/index.mjs";
+export * from "./profiles/index.mjs";
+export * from "./fetchData.mjs";
