+++ src/js/api/posts/index.mjs
@@ -1,9 +1,5 @@
-export * from "./fetchData.mjs";
 export * from "./create.mjs";
 export * from "./postComment.mjs";
 export * from "./editPost.mjs";
 export * from "./delete.mjs";
-export * from "./editProfile.mjs";
 export * from "./likePost.mjs";
-export * from "./followUser.mjs";
-export * from "./unfollowUser.mjs";
