+++ src/js/api/profiles/index.mjs
@@ -1,0 +1,5 @@
+export * from "./editProfile.mjs";
+export * from "./followUser.mjs";
+export * from "./loginUser.mjs";
+export * from "./unfollowUser.mjs";
+export * from "./registerUser.mjs";
