+++ src/js/listeners/posts/oldToNew.mjs
@@ -17,8 +17,8 @@ export function oldToNewListener() {
       posts.forEach((post) => {
         const card = new Thumbnail(
           post.title,
-          post.body,
           post.media,
+          post.tags,
           post.created,
           post.id,
           post.author.name,
