+++ src/js/listeners/posts/postComment.mjs
@@ -1,4 +1,4 @@
-import * as requests from "../../api/apiRequests/index.mjs";
+import * as requests from "../../api/posts/index.mjs";
 
 /**
  * Listen to submit event on comment form and
