+++ src/js/renders/renderUserSpecificPosts.mjs
@@ -1,7 +1,7 @@
-import { fetchData } from "../api/apiRequests/index.mjs";
+import { fetchData } from "../api/index.mjs";
 import { Thumbnail } from "../Class/index.mjs";
-import { BASE_URL } from "../api/constants.mjs";
-import { errorMessage } from "../templates/errorMessage.mjs";
+import { BASE_URL } from "../api/index.mjs";
+import { errorMessage } from "../templates/index.mjs";
 import { getParam } from "../utils/getParam.mjs";
 
 const container = document.querySelector("#user-specific-posts-container");
